+++ tests/testthat/test-class-PKNCAresults.R
@@ -39,23 +39,22 @@ test_that("PKNCAresults generation", {
   verify.result <-
     tibble::tibble(
       treatment="Trt 1",
-      ID=as.integer(rep(c(1, 2), each=14)),
+      ID=as.integer(rep(c(1, 2), each=15)),
       start=0,
-      end=c(24, rep(Inf, 13),
-            24, rep(Inf, 13)),
+      end=c(24, rep(Inf, 14),
+            24, rep(Inf, 14)),
       PPTESTCD=rep(c("auclast", "cmax", "tmax", "tlast", "clast.obs",
                      "lambda.z", "r.squared", "adj.r.squared",
-                     "lambda.z.time.first", "lambda.z.n.points",
-                     "clast.pred", "half.life", "span.ratio",
-                     "aucinf.obs"),
+                     "lambda.z.time.first", "lambda.z.time.last", "lambda.z.n.points",
+                     "clast.pred", "half.life", "span.ratio", "aucinf.obs"),
                    times=2),
       PPORRES=c(13.54, 0.9998, 4.000, 24.00, 0.3441,
-                0.04297, 0.9072, 0.9021, 5.000,
+                0.04297, 0.9072, 0.9021, 5.000, 24.00,
                 20.00, 0.3356, 16.13, 1.178,
                 21.55, 14.03, 0.9410, 2.000,
-                24.00, 0.3148, 0.05689, 0.9000, 0.8944,
-                5.000, 20.00, 0.3011, 12.18,
-                1.560, 19.56),
+                24.00, 0.3148, 0.05689, 0.9000,
+                0.8944, 5.000, 24.00, 20.00, 0.3011,
+                12.18, 1.560, 19.56),
       exclude=NA_character_
     )
   expect_equal(
@@ -126,7 +125,8 @@ test_that("PKNCAresults has exclude, when applicable", {
           c(
             "adj.r.squared", "aucinf.obs", "auclast", "clast.obs",
             "clast.pred", "cmax", "half.life", "lambda.z", "lambda.z.n.points",
-            "lambda.z.time.first", "r.squared", "span.ratio", "tlast", "tmax"
+            "lambda.z.time.first", "lambda.z.time.last", "r.squared",
+            "span.ratio", "tlast", "tmax"
           )
     ),
     info="verify that only expected results are present"
@@ -149,7 +149,7 @@ test_that("PKNCAresults has exclude, when applicable", {
         !(o_result_df$ID == 2 &
             o_result_df$PPTESTCD %in%
             c("lambda.z", "r.squared", "adj.r.squared", "lambda.z.time.first",
-              "lambda.z.n.points", "clast.pred", "half.life", "span.ratio")
+              "lambda.z.time.last", "lambda.z.n.points", "clast.pred", "half.life", "span.ratio")
         )
         ]
     ),
@@ -396,6 +396,10 @@ test_that("getGroups.PKNCAresults", {
   expect_equal(
     getGroups(o_result, level=2:3),
     o_result$result[, c("ID", "start")]
+  )
+  expect_equal(
+    getGroups(myresult, level=3:4),
+    myresult$result[, c("start", "end")]
   )
 })